fix: quick refactor of the authentication into more digestible files
